# teams/src/main/resources/openapi/openapi.yml
openapi: 3.0.3
info:
  title: Teams microservice REST API Docs
  description: TeamUp Teams microservice
  version: 1.0.0
servers:
  - url: http://localhost:8081
    description: Local development server
  - url: https://api.example.com
    description: Production server
...
paths:
...
  /api/v1/teams:
    get:
      tags:
        - Teams
      summary: List Teams REST API
      description: Get all teams with pagination. Optionally filter by userId to get teams where the user is a member (my teams).
      operationId: listTeams
      parameters:
        - name: page
          in: query
          required: false
          schema:
            type: integer
            default: 0
          description: Page number (0-indexed)
        - name: size
          in: query
          required: false
          schema:
            type: integer
            default: 10
          description: Page size
        - name: userId
          in: query
          required: false
          schema:
            type: integer
            format: int64
          description: When set, return only teams where this user is a member (my teams)
      responses:
        '200':
          description: HTTP Status OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PagingTeamResponseDto'
  /api/v1/teams/create:
    post:
      tags:
        - Teams
      summary: Create Team REST API
      description: Submit a request to create a team/project
      operationId: createTeam
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TeamRequestDto'
      responses:
        '201':
          description: HTTP Status CREATED
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResponseDto'
  /api/v1/teams/join:
    post:
      tags:
        - Teams
      summary: Join Team REST API
      description: Submit a request to join a team/project
      operationId: joinTeam
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TeamMemberRequestDto'
      responses:
        '201':
          description: HTTP Status CREATED
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResponseDto'
  /api/v1/teams/{cardId}/invite:
    post:
      tags:
        - Teams
      summary: Invite to Team REST API
      description: Invite a user to join a team/project
      operationId: inviteToTeam
      parameters:
        - name: cardId
          in: path
          required: true
          schema:
            type: integer
            format: int64
          description: Card ID
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TeamMemberRequestDto'
      responses:
        '201':
          description: HTTP Status CREATED
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResponseDto'
  /api/v1/teams/{cardId}/status:
    put:
      tags:
        - Teams
      summary: Update Member Status REST API
      description: Update the status of a team member (accept/reject)
      operationId: updateMemberStatus
      parameters:
        - name: cardId
          in: path
          required: true
          schema:
            type: integer
            format: int64
          description: Card ID
        - name: userId
          in: query
          required: true
          schema:
            type: integer
            format: int64
          description: User ID
        - name: status
          in: query
          required: true
          schema:
            type: string
          description: New status
      responses:
        '200':
          description: HTTP Status OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TeamMemberResponseDto'
        '404':
          description: Team member not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResponseDto'
  /api/v1/teams/{cardId}/remove:
    delete:
      tags:
        - Teams
      summary: Remove Team Member REST API
      description: Remove a user from a team/project
      operationId: removeMember
      parameters:
        - name: cardId
          in: path
          required: true
          schema:
            type: integer
            format: int64
          description: Card ID
        - name: userId
          in: query
          required: true
          schema:
            type: integer
            format: int64
          description: User ID
      responses:
        '200':
          description: HTTP Status OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResponseDto'
        '404':
          description: Team member not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResponseDto'
        '500':
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResponseDto'
  /api/v1/teams/{cardId}:
    get:
      tags:
        - Teams
      summary: Get Team Members REST API
      description: Get all members of a specific team/project with pagination
      operationId: getTeamMembers
      parameters:
        - name: cardId
          in: path
          required: true
          schema:
            type: integer
            format: int64
          description: Card ID
        - name: page
          in: query
          required: false
          schema:
            type: integer
            default: 0
          description: Page number (0-indexed)
        - name: size
          in: query
          required: false
          schema:
            type: integer
            default: 10
          description: Page size
      responses:
        '200':
          description: HTTP Status OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PagingTeamMemberResponseDto'
  /api/v1/teams/fetch:
    get:
      tags:
        - Teams
      summary: Get Team Info REST API
      description: Get info of a specific team
      operationId: fetch
      parameters:
        - name: teamId
          in: query
          required: true
          schema:
            type: integer
            format: int64
          description: Team ID
      responses:
        '200':
          description: HTTP Status OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TeamResponseDto'
  /api/v1/teams/{teamId}/members:
    get:
      tags:
        - Teams
      summary: Get Team Members by Team ID REST API
      description: Get all members of a team with pagination
      operationId: getTeamMembersByTeamId
      parameters:
        - name: teamId
          in: path
          required: true
          schema:
            type: integer
            format: int64
          description: Team ID
        - name: page
          in: query
          required: false
          schema:
            type: integer
            default: 0
          description: Page number (0-indexed)
        - name: size
          in: query
          required: false
          schema:
            type: integer
            default: 10
          description: Page size
      responses:
        '200':
          description: HTTP Status OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PagingTeamMemberResponseDto'
components:
  schemas:
...
    ResponseDto:
      type: object
      properties:
        statusCode:
          type: string
          description: Status code of the response
          example: "200"
        statusMessage:
          type: string
          description: Status message of the response
          example: "Request processed successfully"
    TeamRequestDto:
      type: object
      required:
        - name
      properties:
        teamId:
          type: integer
          format: int64
          description: ID of the team
          example: 1
        name:
          type: string
          description: Name of the team
          example: "Super cool team"
        description:
          type: string
          description: Description of the team
          example: "Description of team"
    TeamResponseDto:
      type: object
      properties:
        id:
          type: integer
          format: int64
          description: ID of the team
          example: 1
        name:
          type: string
          description: Name of the team
          example: "Test Name"
        description:
          type: string
          description: Description of the team
          example: "Test Description"
        teamMembers:
          type: array
          description: Team Members of the team
          items:
            $ref: '#/components/schemas/TeamMemberResponseDto'
    TeamMemberRequestDto:
      type: object
      required:
        - userId
        - role
        - teamId
      properties:
        userId:
          type: integer
          format: int64
          description: ID of the user to add to the team
          example: 1
        role:
          type: string
          enum:
            - CREATOR
            - PARTICIPANT
            - LEADER
          description: Role of the user in the team
          example: "PARTICIPANT"
        teamId:
          type: integer
          format: int64
          description: Team Id where user is added
          example: 1
        cardId:
          type: integer
          format: int64
          description: Card Id from which user added himself
          example: 1
    TeamMemberResponseDto:
      type: object
      properties:
        id:
          type: integer
          format: int64
          description: Team member identifier
          example: 1
        cardId:
          type: integer
          format: int64
          description: ID of the card/project
          example: 1
        teamId:
          type: integer
          format: int64
          description: ID of the team
          example: 1
        userId:
          type: integer
          format: int64
          description: ID of the user
          example: 1
        role:
          type: string
          description: Role of the user in the team
          example: "PARTICIPANT"
        status:
          type: string
          description: Status of participation
          example: "JOINED"
        joinedAt:
          type: string
          format: date-time
          description: Date and time when the user joined the team
          example: "2023-01-01T10:00:00"
        createdAt:
          type: string
          format: date-time
          description: Date and time when the team member was created
          example: "2023-01-01T10:00:00"
        createdBy:
          type: string
          description: User who created the team member
          example: "admin"
        updatedAt:
          type: string
          format: date-time
          description: Date and time when the team member was last updated
          example: "2023-01-02T11:00:00"
        updatedBy:
          type: string
          description: User who last updated the team member
          example: "admin"
    PagingDto:
      type: object
      properties:
        totalElements:
          type: integer
          format: int64
          description: Total number of elements across all pages
          example: 100
        totalPages:
          type: integer
          description: Total number of pages
          example: 10
        number:
          type: integer
          description: Current page number (0-indexed)
          example: 0
        size:
          type: integer
          description: Number of elements per page
          example: 10
...
    PagingTeamMemberResponseDto:
      allOf:
        - $ref: '#/components/schemas/PagingDto'
        - type: object
          properties:
            content:
              type: array
              description: List of team members in the current page
              items:
                $ref: '#/components/schemas/TeamMemberResponseDto'
    PagingTeamResponseDto:
      allOf:
        - $ref: '#/components/schemas/PagingDto'
        - type: object
          properties:
            content:
              type: array
              description: List of teams in the current page
              items:
                $ref: '#/components/schemas/TeamResponseDto'